[New] Added test cases for pages admin users form
